+++ Sources/Sentry/SentryStacktraceBuilder.m
@@ -24,6 +24,7 @@ SentryStacktraceBuilder ()
 {
     if (self = [super init]) {
         self.crashStackEntryMapper = crashStackEntryMapper;
+        self.symbolicate = NO;
     }
     return self;
 }
@@ -40,7 +41,7 @@ SentryStacktraceBuilder ()
             // skip the marker frame
             continue;
         }
-        if (stackCursor.symbolicate(&stackCursor)) {
+        if (self.symbolicate == false || stackCursor.symbolicate(&stackCursor)) {
             frame = [self.crashStackEntryMapper mapStackEntryWithCursor:stackCursor];
             [frames addObject:frame];
         }