+++ Tests/SentryTests/SentryCrash/SentryBinaryImageCacheTests.m
@@ -1,6 +1,8 @@
+#import "SentryBinaryImageCache+Private.h"
 #import "SentryCrashBinaryImageCache.h"
 #import "SentryCrashWrapper.h"
 #import <XCTest/XCTest.h>
+
 #include <mach-o/dyld.h>
 
 // Exposing test only functions from `SentryCrashBinaryImageCache.m`
@@ -68,11 +70,11 @@ delayAddBinaryImage(void)
     }
 }
 
-@interface SentryBinaryImageCacheTests : XCTestCase
+@interface SentryCrashBinaryImageCacheTests : XCTestCase
 
 @end
 
-@implementation SentryBinaryImageCacheTests
+@implementation SentryCrashBinaryImageCacheTests
 
 + (void)setUp
 {
@@ -251,6 +253,31 @@ delayAddBinaryImage(void)
     dispatch_semaphore_signal(delaySemaphore);
     [self assertBinaryImageCacheLength:0];
     XCTAssertEqual(result, 0);
+}
+
+// Adding a SentryBinaryImageCache test inside
+// SentryCrashBinaryImageCache to test integration between both
+// because is easier to control SentryCrashBinaryImageCache in an objc test
+- (void)testSentryBinaryImageCacheIntegration
+{
+    sentrycrashbic_startCache();
+
+    SentryBinaryImageCache *imageCache = SentryBinaryImageCache.shared;
+    [imageCache start];
+    // by calling start, SentryBinaryImageCache will register a callback with
+    // `SentryCrashBinaryImageCache` that should be called for every image already cached.
+    XCTAssertEqual(5, imageCache.cache.count);
+
+    addBinaryImage([mach_headers_test_cache[5] pointerValue], 0);
+    XCTAssertEqual(6, imageCache.cache.count);
+
+    removeBinaryImage([mach_headers_expect_array[1] pointerValue], 0);
+    removeBinaryImage([mach_headers_expect_array[2] pointerValue], 0);
+    XCTAssertEqual(4, imageCache.cache.count);
+    [imageCache stop];
+
+    addBinaryImage([mach_headers_test_cache[6] pointerValue], 0);
+    XCTAssertNil(imageCache.cache);
 }
 
 - (void)assertBinaryImageCacheLength:(int)expected