+++ Tests/SentryTests/SentryCrash/SentryThreadInspectorTests.swift
@@ -1,17 +1,20 @@
+import SentryTestUtils
 import XCTest
 
 class SentryThreadInspectorTests: XCTestCase {
     
     private class Fixture {
         var testMachineContextWrapper = TestMachineContextWrapper()
-        var stacktraceBuilder = TestSentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: [])))
+        var stacktraceBuilder = TestSentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: []), binaryImageCache: SentryBinaryImageCache.shared ))
         var keepThreadAlive = true
         
-        func getSut(testWithRealMachineContextWrapper: Bool = false) -> SentryThreadInspector {
+        func getSut(testWithRealMachineContextWrapper: Bool = false, symbolicate: Bool = true) -> SentryThreadInspector {
             
             let machineContextWrapper = testWithRealMachineContextWrapper ? SentryCrashDefaultMachineContextWrapper() : testMachineContextWrapper as SentryCrashMachineContextWrapper
-            let stacktraceBuilder = testWithRealMachineContextWrapper ? SentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: []))) : self.stacktraceBuilder
-            
+            let stacktraceBuilder = testWithRealMachineContextWrapper ? SentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: []), binaryImageCache: SentryBinaryImageCache.shared)) : self.stacktraceBuilder
+
+            stacktraceBuilder.symbolicate = symbolicate
+
             return SentryThreadInspector(
                 stacktraceBuilder: stacktraceBuilder,
                 andMachineContextWrapper: machineContextWrapper
@@ -24,6 +27,11 @@ class SentryThreadInspectorTests: XCTestCase {
     override  func setUp() {
         super.setUp()
         fixture = Fixture()
+    }
+
+    override class func tearDown() {
+        super.tearDown()
+        clearTestState()
     }
     
     func testNoThreads() {
@@ -109,6 +117,17 @@ class SentryThreadInspectorTests: XCTestCase {
         XCTAssertGreaterThan(stackTrace.frames.count, 0)
         XCTAssertNotEqual(stackTrace.frames.first?.instructionAddress, "0x0000000000000000")
         XCTAssertNotEqual(stackTrace.frames.first?.function, "<redacted>")
+    }
+
+    func testStackTrackForCurrentThreadAsyncUnsafe_NoSymbolication() {
+        guard let stackTrace = fixture.getSut(testWithRealMachineContextWrapper: true, symbolicate: false).stacktraceForCurrentThreadAsyncUnsafe() else {
+            XCTFail("Stack Trace not found")
+            return
+        }
+
+        XCTAssertNotNil(stackTrace)
+        XCTAssertGreaterThan(stackTrace.frames.count, 0)
+        XCTAssertEqual(stackTrace.frames.first?.function, "<redacted>")
     }
 
     func testOnlyCurrentThreadHasStacktrace() {